backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-dotenv>=1.0.0
spotipy>=2.23.0
requests>=2.31.0
beautifulsoup4>=4.12.2
pyperclip>=1.8.2